refactor(test): 테스트 코드에서 DDL 실행과 커넥션 관리를 추상화 및 적용
